+++ .changeset/thirty-bats-notice.md
@@ -1,0 +1,5 @@
+---
+'@mastra/schema-compat': patch
+---
+
+Add support for zod v4
