+++ packages/schema-compat/src/provider-compats/anthropic.ts
@@ -1,8 +1,12 @@
-import type { ZodType as ZodTypeV3 } from 'zod/v3';
-import type { ZodType as ZodTypeV4 } from 'zod/v4';
+import { z } from 'zod';
+import type { ZodType as ZodTypeV3, ZodObject as ZodObjectV3 } from 'zod/v3';
+import type { ZodType as ZodTypeV4, ZodObject as ZodObjectV4 } from 'zod/v4';
 import type { Targets } from 'zod-to-json-schema';
 import { SchemaCompatLayer } from '../schema-compatibility';
+import type { AllZodType as AllZodTypeV3 } from '../schema-compatibility-v3';
+import type { AllZodType as AllZodTypeV4 } from '../schema-compatibility-v4';
 import type { ModelInformation } from '../types';
+import { isOptional, isObj, isArr, isUnion, isString } from '../zodTypes';
 
 export class AnthropicSchemaCompatLayer extends SchemaCompatLayer {
   constructor(model: ModelInformation) {
@@ -20,24 +24,24 @@ export class AnthropicSchemaCompatLayer extends SchemaCompatLayer {
   processZodType(value: ZodTypeV3): ZodTypeV3;
   processZodType(value: ZodTypeV4): ZodTypeV4;
   processZodType(value: ZodTypeV3 | ZodTypeV4): ZodTypeV3 | ZodTypeV4 {
-    if (this.isOptional(value)) {
-      const handleTypes = ['ZodObject', 'ZodArray', 'ZodUnion', 'ZodNever', 'ZodUndefined', 'ZodTuple'];
+    if (isOptional(z)(value)) {
+      const handleTypes: AllZodTypeV3[] | AllZodTypeV4 = [
+        'ZodObject',
+        'ZodArray',
+        'ZodUnion',
+        'ZodNever',
+        'ZodUndefined',
+        'ZodTuple',
+      ];
       if (this.getModel().modelId.includes('claude-3.5-haiku')) handleTypes.push('ZodString');
-      // @ts-expect-error - fix later
       return this.defaultZodOptionalHandler(value, handleTypes);
-      // @ts-expect-error - fix later
-    } else if (this.isObj(value)) {
-      // @ts-expect-error - fix later
+    } else if (isObj(z)(value)) {
       return this.defaultZodObjectHandler(value);
-    } else if (this.isArr(value)) {
-      // @ts-expect-error - fix later
+    } else if (isArr(z)(value)) {
       return this.defaultZodArrayHandler(value, []);
-      // @ts-expect-error - fix later
-    } else if (this.isUnion(value)) {
-      // @ts-expect-error - fix later
+    } else if (isUnion(z)(value)) {
       return this.defaultZodUnionHandler(value);
-      // @ts-expect-error - fix later
-    } else if (this.isString(value)) {
+    } else if (isString(z)(value)) {
       // the claude-3.5-haiku model support these properties but the model doesn't respect them, but it respects them when they're
       // added to the tool description
 
@@ -48,6 +52,10 @@ export class AnthropicSchemaCompatLayer extends SchemaCompatLayer {
       }
     }
 
-    return this.defaultUnsupportedZodTypeHandler(value, ['ZodNever', 'ZodTuple', 'ZodUndefined']);
+    return this.defaultUnsupportedZodTypeHandler(value as ZodObjectV4<any> | ZodObjectV3<any>, [
+      'ZodNever',
+      'ZodTuple',
+      'ZodUndefined',
+    ]);
   }
 }
