+++ packages/schema-compat/src/provider-compats/google.ts
@@ -1,9 +1,10 @@
 import { z } from 'zod';
-import type { ZodType as ZodTypeV3 } from 'zod/v3';
-import type { ZodType as ZodTypeV4 } from 'zod/v4';
+import type { ZodType as ZodTypeV3, ZodObject as ZodObjectV3 } from 'zod/v3';
+import type { ZodType as ZodTypeV4, ZodObject as ZodObjectV4 } from 'zod/v4';
 import type { Targets } from 'zod-to-json-schema';
 import { SchemaCompatLayer } from '../schema-compatibility';
 import type { ModelInformation } from '../types';
+import { isOptional, isNull, isObj, isArr, isUnion, isString, isNumber } from '../zodTypes';
 
 export class GoogleSchemaCompatLayer extends SchemaCompatLayer {
   constructor(model: ModelInformation) {
@@ -20,42 +21,29 @@ export class GoogleSchemaCompatLayer extends SchemaCompatLayer {
   processZodType(value: ZodTypeV3): ZodTypeV3;
   processZodType(value: ZodTypeV4): ZodTypeV4;
   processZodType(value: ZodTypeV3 | ZodTypeV4): ZodTypeV3 | ZodTypeV4 {
-    if (this.isOptional(value as ZodTypeV4)) {
-      // @ts-expect-error - fix later
-      return this.defaultZodOptionalHandler(value, [
-        'ZodObject',
-        'ZodArray',
-        'ZodUnion',
-        'ZodString',
-        'ZodNumber',
-        // @ts-expect-error - fix later
-        ...this.getUnsupportedZodTypes(value),
-      ]);
-    } else if (this.isNull(value as ZodTypeV4)) {
+    if (isOptional(z)(value)) {
+      return this.defaultZodOptionalHandler(value, ['ZodObject', 'ZodArray', 'ZodUnion', 'ZodString', 'ZodNumber']);
+    } else if (isNull(z)(value)) {
       // Google models don't support null, so we need to convert it to any and then refine it to null
       return z
         .any()
         .refine(v => v === null, { message: 'must be null' })
         .describe(value.description || 'must be null');
-    } else if (this.isObj(value as ZodTypeV4)) {
-      // @ts-expect-error - fix later
+    } else if (isObj(z)(value)) {
       return this.defaultZodObjectHandler(value);
-    } else if (this.isArr(value as ZodTypeV4)) {
-      // @ts-expect-error - fix later
+    } else if (isArr(z)(value)) {
       return this.defaultZodArrayHandler(value, []);
-    } else if (this.isUnion(value as ZodTypeV4)) {
-      // @ts-expect-error - fix later
+    } else if (isUnion(z)(value)) {
       return this.defaultZodUnionHandler(value);
-    } else if (this.isString(value as ZodTypeV4)) {
+    } else if (isString(z)(value)) {
       // Google models support these properties but the model doesn't respect them, but it respects them when they're
       // added to the tool description
-      return this.defaultZodStringHandler(value as ZodStringV4);
-    } else if (this.isNumber(value as ZodTypeV4)) {
+      return this.defaultZodStringHandler(value);
+    } else if (isNumber(z)(value)) {
       // Google models support these properties but the model doesn't respect them, but it respects them when they're
       // added to the tool description
       return this.defaultZodNumberHandler(value);
     }
-    // @ts-expect-error - fix later
-    return this.defaultUnsupportedZodTypeHandler(value);
+    return this.defaultUnsupportedZodTypeHandler(value as ZodObjectV4<any> | ZodObjectV3<any>);
   }
 }
