+++ packages/schema-compat/src/provider-compats/openai-reasoning.ts
@@ -1,12 +1,12 @@
 import { z } from 'zod';
-import type { ZodType as ZodTypeV3 } from 'zod/v3';
-import type { ZodType as ZodTypeV4 } from 'zod/v4';
+import type { ZodType as ZodTypeV3, ZodObject as ZodObjectV3 } from 'zod/v3';
+import type { ZodType as ZodTypeV4, ZodObject as ZodObjectV4 } from 'zod/v4';
 import type { Targets } from 'zod-to-json-schema';
 import { SchemaCompatLayer } from '../schema-compatibility';
 import type { ModelInformation } from '../types';
+import { isOptional, isObj, isArr, isUnion, isDefault, isNumber, isString, isDate } from '../zodTypes';
 
 export class OpenAIReasoningSchemaCompatLayer extends SchemaCompatLayer {
-  
   constructor(model: ModelInformation) {
     super(model);
   }
@@ -39,16 +39,16 @@ export class OpenAIReasoningSchemaCompatLayer extends SchemaCompatLayer {
   processZodType(value: ZodTypeV3): ZodTypeV3;
   processZodType(value: ZodTypeV4): ZodTypeV4;
   processZodType(value: ZodTypeV3 | ZodTypeV4): ZodTypeV3 | ZodTypeV4 {
-    if (this.isOptional(value)) {
+    if (isOptional(z)(value)) {
       const innerZodType = this.processZodType(value._def.innerType);
       return innerZodType.nullable();
-    } else if (this.isObj(value)) {
+    } else if (isObj(z)(value)) {
       return this.defaultZodObjectHandler(value, { passthrough: false });
-    } else if (this.isArr(value)) {
+    } else if (isArr(z)(value)) {
       return this.defaultZodArrayHandler(value);
-    } else if (this.isUnion(value)) {
+    } else if (isUnion(z)(value)) {
       return this.defaultZodUnionHandler(value);
-    } else if (this.isDefault(value)) {
+    } else if (isDefault(z)(value)) {
       const defaultDef = value._def;
       const innerType = defaultDef.innerType;
       const defaultValue = defaultDef.defaultValue();
@@ -63,13 +63,13 @@ export class OpenAIReasoningSchemaCompatLayer extends SchemaCompatLayer {
         result = result.describe(description);
       }
       return result;
-    } else if (this.isNumber(value)) {
+    } else if (isNumber(z)(value)) {
       return this.defaultZodNumberHandler(value);
-    } else if (this.isString(value)) {
+    } else if (isString(z)(value)) {
       return this.defaultZodStringHandler(value);
-    } else if (this.isDate(value)) {
+    } else if (isDate(z)(value)) {
       return this.defaultZodDateHandler(value);
-    } else if (value._def.typeName === 'ZodAny') {
+    } else if (value.constructor.name === 'ZodAny') {
       // It's bad practice in the tool to use any, it's not reasonable for models that don't support that OOTB, to cast every single possible type
       // in the schema. Usually when it's "any" it could be a json object or a union of specific types.
       return z
@@ -80,6 +80,6 @@ export class OpenAIReasoningSchemaCompatLayer extends SchemaCompatLayer {
         );
     }
 
-    return this.defaultUnsupportedZodTypeHandler(value);
+    return this.defaultUnsupportedZodTypeHandler(value as ZodObjectV4<any> | ZodObjectV3<any>);
   }
 }
