+++ packages/schema-compat/src/provider-compats/openai.ts
@@ -1,8 +1,10 @@
-import type { ZodType as ZodTypeV3 } from 'zod/v3';
-import type { ZodType as ZodTypeV4, ZodString as ZodStringV4 } from 'zod/v4';
+import { z } from 'zod';
+import type { ZodType as ZodTypeV3, ZodObject as ZodObjectV3 } from 'zod/v3';
+import type { ZodType as ZodTypeV4, ZodObject as ZodObjectV4 } from 'zod/v4';
 import type { Targets } from 'zod-to-json-schema';
 import { SchemaCompatLayer } from '../schema-compatibility';
 import type { ModelInformation } from '../types';
+import { isOptional, isObj, isUnion, isArr, isString } from '../zodTypes';
 
 export class OpenAISchemaCompatLayer extends SchemaCompatLayer {
   constructor(model: ModelInformation) {
@@ -27,7 +29,7 @@ export class OpenAISchemaCompatLayer extends SchemaCompatLayer {
   processZodType(value: ZodTypeV3): ZodTypeV3;
   processZodType(value: ZodTypeV4): ZodTypeV4;
   processZodType(value: ZodTypeV3 | ZodTypeV4): ZodTypeV3 | ZodTypeV4 {
-    if (this.isOptional(value as ZodTypeV4)) {
+    if (isOptional(z)(value)) {
       return this.defaultZodOptionalHandler(value, [
         'ZodObject',
         'ZodArray',
@@ -37,25 +39,27 @@ export class OpenAISchemaCompatLayer extends SchemaCompatLayer {
         'ZodUndefined',
         'ZodTuple',
       ]);
-    } else if (this.isObj(value as ZodTypeV4)) {
+    } else if (isObj(z)(value)) {
       return this.defaultZodObjectHandler(value);
-    } else if (this.isUnion(value as ZodTypeV4)) {
+    } else if (isUnion(z)(value)) {
       return this.defaultZodUnionHandler(value);
-    } else if (this.isArr(value as ZodTypeV4)) {
+    } else if (isArr(z)(value)) {
       return this.defaultZodArrayHandler(value);
-    } else if (this.isString(value as ZodTypeV4)) {
+    } else if (isString(z)(value)) {
       const model = this.getModel();
-      const checks: string[] = ['emoji'];
+      const checks = ['emoji'] as const;
 
       if (model.modelId.includes('gpt-4o-mini')) {
-        checks.push('regex');
+        return this.defaultZodStringHandler(value, ['emoji', 'regex']);
       }
 
-      // @ts-expect-error - fix later
       return this.defaultZodStringHandler(value, checks);
     }
 
-    // @ts-expect-error - fix later
-    return this.defaultUnsupportedZodTypeHandler(value, ['ZodNever', 'ZodUndefined', 'ZodTuple']);
+    return this.defaultUnsupportedZodTypeHandler(value as ZodObjectV4<any> | ZodObjectV3<any>, [
+      'ZodNever',
+      'ZodUndefined',
+      'ZodTuple',
+    ]);
   }
 }
