+++ packages/schema-compat/src/utils-v3.test.ts
@@ -1,0 +1,9 @@
+import { vi } from 'vitest';
+import { z as zV3 } from 'zod/v3';
+import { runTestSuite } from './utils-test-suite';
+
+vi.mock('zod', () => ({
+  z: zV3,
+}));
+
+runTestSuite();
