+++ packages/schema-compat/src/utils-v4.test.ts
@@ -1,0 +1,9 @@
+import { vi } from 'vitest';
+import { z as zV4 } from 'zod/v4';
+import { runTestSuite } from './utils-test-suite';
+
+vi.mock('zod', () => ({
+  z: zV4,
+}));
+
+runTestSuite();
