+++ packages/schema-compat/src/zod-to-json.ts
@@ -1,5 +1,5 @@
 import type { JSONSchema7 } from 'json-schema';
-import z from 'zod';
+import { z } from 'zod';
 import type { ZodSchema as ZodSchemaV3 } from 'zod/v3';
 import type { ZodType as ZodSchemaV4 } from 'zod/v4';
 import type { Targets } from 'zod-to-json-schema';
